Split models into multiple files
